// packages/sdk/src/generated/operations.ts
export const operations = {
  "activeDevice": {
    operationId: "activeDevice",
    method: "post",
    path: "/openapi/v1/{omadacId}/sites/{siteId}/multi-devices/active",
    tags: ["Device"],
    summary: "Batch active devices",
    deprecated: false,
  },
// ...
} as const satisfies Record<string, OperationInfo>;

export type OperationId = keyof typeof operations;